+++ package/kvs-in-memory/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @flex/kvs-in-memory
 
+## 1.1.0
+
+### Minor Changes
+
+- Extends KVS interface
+- Implements the In-Memory Storage
+
 ## 1.0.0
 
 ### Major Changes
@@ -9,5 +16,5 @@
 ### Patch Changes
 
 - Updated dependencies
-  - @flex/base-error@1.0.0
-  - @flex/kvs-core@1.0.0
+    - @flex/base-error@1.0.0
+    - @flex/kvs-core@1.0.0